# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.OrderModel:
    properties:
      created_at:
        type: string
      id:
        type: string
      order_date:
        type: string
      order_id:
        type: string
      table_id:
        type: string
      updated_at:
        type: string
    required:
    - order_date
    - table_id
    type: object
...
host: localhost:3000
info:
  contact:
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive Restaurant Management System API for managing users,
    menus, food items, orders, tables, invoices, and notes. Available on both local
    and online servers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Restaurant Management API
  version: "1.0"
  x-servers:
  - description: Local Development Server
    url: http://localhost:3000/api/v1
  - description: Online Production Server
    url: https://restaurant-management-f9kx.onrender.com/api/v1
paths:
...
  /order:
    post:
      consumes:
      - application/json
      description: Create a new order for a table
      parameters:
      - description: Order data (table_id required)
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.OrderModel'
      produces:
      - application/json
      responses:
        "200":
          description: Order created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new order
      tags:
      - Order
  /order/{id}:
    delete:
      description: Delete an order by order ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an order
      tags:
      - Order
    get:
      description: Retrieve order details by order ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/model.OrderModel'
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a specific order
      tags:
      - Order
    put:
      consumes:
      - application/json
      description: Update order details by order ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated order data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/model.OrderModel'
      produces:
      - application/json
      responses:
        "200":
          description: Order updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an order
      tags:
      - Order
...
  /orders:
    get:
      description: Retrieve a list of all orders
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            items:
              $ref: '#/definitions/model.OrderModel'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all orders
      tags:
      - Order
...
schemes:
- http
- https
swagger: "2.0"